January 5th, 2020, 9:40pm - 10:57pm
(For the work session on December 16th, 2020, I meant to write 9:40 pm - 11:05 pm rather than 10:05 pm) 
In this work session, I finished addressing another two problems. I spent a lot of time on first question to figure out the calculation of dominant offsprings. It took me a while to realize if both parents are homozygous dominant, then all of the possible offsprings would be 100% dominant. If one of the parent is homozygous dominant, then the number of offsprings containing a dominant allele equals to the number of homozygous parents times the other parents. After figuring out the calculation, the coding was not hard. I figured out to use combination to make my calculation easier.  For the second question, it took me a lot of time to write the codes. I had to search for the use of "strip" to finish the question. I also need to learn to use "open" and "read" to read the lines in the file provided by the question. I used a dictionary in order to return two corresponding results.