January 10th, 2020, 6:20pm - 7:50pm
In this working session, I finished another two questions on Rosalind. The first question asks to find locations of motifs in a DNA sequence. I first tried to use regular expression operations (RE) to solve this problem and did some research on the internet, but then I realized it might not be the most effective solution. I then came up with the current code. One thing to notice is that in this question, the index for the first element in a string is "1" rather than "0". While doing the second question, I spent a lot of time trying to figure out recurrence relation mentioned in the question. At first I did not realize that small rabbits need to grow to reproduction age to reproduce more rabbits, so I comprehended the question wrong and came up with the wrong solution. Understanding the question itself requires me to go on the internet and search for the logics behind rabbit reproduction and recurrence relations. As soon as I figured out what the question is saying, the coding part is not hard.