January 13th, 2020, 10:00pm - 11:35pm
In this working session, I finished a question asking to give all the possible permutations with numbers smaller than or equal to a given number. I solved this problem with multiple functions rather than one big function. I first figured out a way to calculate the factorial of the number, because the factorial gives the total number of all possible permutations. I then managed to write a function that can visualize all the permutations. It took me a while to recall how permutations work and to think of a way to exhibit all the possible permutations. I started another question asking to find the longest common substring among multiple DNA sequences. I currently do not have a solution to this problem. My current idea is to first find the shortest sequence because it is the longest possible common substring. Then check every element in the shortest sequence to see which parts are shown in other sequences. If there are multiple common sequences, only give out the longest one. I will continue working on this question for the next working session.