January 17th, 2020, 1:40am - 2:58am
In this working session, I continued working on the question left over during last working session. I could not figure out a way to implement my original idea of finding the shortest sequence, and then check if the shortest sequence is contained by other sequences. If not, eliminate a nucleotide from the shortest sequence and redo the comparison process until the common motif appears. However, I did not find a way to write the code, so I thought about a new algorithm method. This method breaks DNA sequence into fragments, and compare whether each fragment appears in all other sequences. After all, extract the longest common fragment to be the common motif. I did not know how to extract the longest common fragment out of all the common fragments, so I planned to manually check which one is the longest and submit the longest fragment as the final result. However, a current problem with my code is on line 234 where "segments = fragments(seq[0])", anaconda reports "IndexError: list index out of range", so the function currently does not work. I will continue working on this question for the next working session.