January 21st, 2021, 11:50am - 1:50am
In this working session, I finally figured out the code to solve question 11. I spent a lot of time trying to fix the out of range problem in my previous code, but unfortunately, I could not find a way to fix it. I started to think whether the idea of breaking DNA sequences into fragments is wrong, so I rethought about my idea of selecting the shortest DNA sequence in the working session previous to that working session. This time, I did not ask my code to read the provided file, but copied and pasted the elements in the file into my code to solve the problem more directly. I believed this helped me to write my codes more effectively as I did not need to deal with readlines() or open(). As a result, I can find the shortest DNA sequence and compared it with all other sequences to see if there are common motifs. I managed to use bool to check whether there are common motifs among all the sequences, and I believed it was effective. This question 11 took me two and a half working sessions to finish. It was hard, but I believed I have learned a lot from this question.